feat: explain redirect of unauthed users to login pages
Motivation: explain to unauthenticated users why they have been
redirected to login page by displaying a "You must be logged in to view
this page message".

Implementation: add a SignIn middleware around protected CRUD model
routes that redirects unauthenticated requests to a new sign-in routes.

I'll add a better user-facing message and layout in the future; the
point for now is to implement the backend infrastructure.