feat: switch to placeholder for ingredient_nutrient.amount
Instead of setting ingredient.ingredient_nutrients.*.amount to zero, I
set it to null and use a placeholder value of zero for each number
input (and then convert null amounts to zero in backend). This saves the
user from having to delete the default zero value in frontend. Commit
includes necessary updates to affected services to account for amount
being nullable.