refactor: switch to a single meal_unit
Switches to invariant that meals and food lists have a single "meal/food
list unit" with custom_grams equal to the meal/food list's mass in
grams. It is entirely managed internally and the user cannot interact
with its creation/updating/deletion.

Previously meals supported an arbitrary number of custom units and food
lists did not support custom units at all.

This change simplifies management of meal units and makes it easier to
set default unit in which to display a meal or food list's nutrient
profile.