feat: add Log Data and Trends pages to main nav bars
Adds links to Log Data and Trends pages to the mobile and desktop
navbars used in AppLayout. Slight complication was adding a popover for
Food items on small screens that expands to Ingredients, Meals, and Food
Lists, since otherwise the top navbar becomes too crowded with Log Data
and Trends added.