feat: make it easer to fill in zero values
In Ingredients and IntakeGuidelines CreateOrEdit, resets zero nutrient
amount/rdi values to null, so their input is empty, and the user doesn't
have to delete the zero value before filling in the new value. This is
fine because of the zero placeholder, and because null values are
treated as zero.