feat: add client-side validation to CustomUnitDialog
Adds a bit of home-rolled client-side validation to creation of
ingredient custom units. This reduces the risk of sending an invalid
custom unit to the backend.
The larger goal is to improve the UX when using a newly-created custom
unit as a `ingredient_nutrient_amount_unit`---you want to prevent user
from using a custom unit that would turn out to be invalid after
serverside validation.