feat: add CSV seed files to Git control
Ditto of previous commit. Yes, this does mean we're dumping USDA food,
food_category, and food_nutrient table data into Git control which is
maybe excessive. But it makes it easier to use the same seeding workflow
both locally and in prod, and keeps all the data you need to reconstruct
the app in one Git repo.