feat: add description column to resource tables
Adds a nullable `description` column to main resource tables
(Ingredients, Meals, FoodLists, IntakeGuidelies, BodyWeightRecords, and
food intake records). This is mostly just to have the description for
later without having to change the prod database, since it would be
useful to let user add descriptions to resources for their own records
and I think a description feature is inevitable down the road. Might as
well do it now.