fix: improve Meals/CreateOrEdit input widths on mobile
Previously ingredient input in particular became unusably small on
mobile. This commit fixes this with a few tweaks to SimpleCombobox and
the CreateOrEdit page, and switches to from a table layout for
displaying/editing mealIngredients to a grid layout in the process
(easier to manage widths and more pleasant to interact with).