feat: add optional priority to IntakeGuidelines
This commit adds a `priority` column to the `intake_guidelines` table,
and adds the necessary backend and frontend code to allow users to add
an optional priority to their intake guidelines.

Motivation: give user the choice of which intake guideline is used by
default to compute percent daily value in nutrient profiles. The
guideline with highest priority is used.