revert: only ask user for date_time for Meals/CreateAndLog
Previously I had planned on user inputting meal amount and unit (as per
a normal MealIntakeRecord) in addition to date and time. But in practice
anyone using the createAndLog meals feature will be logging one meal. So
it's simpler for the user and for me, the developer, to remove user
input of meal amount and unit (and just default to 1 meal).