fix: fix autofocus of new ingredient/meal input in CreateOrEdit
In Ingredients/CreateOrEdit and Meals/CreateOrEdit, the goal is to
autofocus the combobox for inputting the new ingredient/meal name on
adding a new ingredient/meal.

I previously had left a broken implementation using Vue refs inside a
v-for from before I had switched from a table to a grid layout. I
replaced this with a much simpler variant using element ids.