wip: default to hidden description input
Most of the time users won't add a description and the text area takes
up a log of space, so instead of showing it by default, I'm adding an
"Add description" button to show the text area.

wip: the change is almost done, I just need to swap the the "Add
description" button below the description label to clearly indicate the
button goes with the description.