wip: slightly improve FuzzyCombobox iOS content jump
In Meals/CreateOrEdit and FoodLists/CreateOrEdit viewport jumps around
as FuzzyCombobox options appear/disappear (doesn't happen on desktop or
in e.g. Brave on iOS). This will have to be addressed rigorously in the
future, but adding some margin to page bottom and decreasing number of
displayed FuzzyCombobox items is a temporary fix with no real downside.