fix: autofocus meal input in LogMealIntakeRecordDialog
Previously, when opening LogMealIntakeRecordDialog for logging new meal
intake records, the "New Meal" button would autofocus (because the
Headless UI Dialog focuses the first focusable item by default). But it
makes more sense to focus meal input, so that the user can begin typing
the meal name right away if so desired (and if they were going to use
the new meal button they realistically be moving their mouse to it and
clicking it anyway, and not taking advantage of the autofocus).