feat: improve mobile layout of Meals/CreateOrEdit
Breaks meal ingredient, amount, and unit inputs into two lines on small
screens---previously these were all on one line, which could not fit in
the viewport and required horizontal scrolling back and forth, which was
an unpleasant user experience.

Includes some refactoring of the grid layout, to move grid layout into
individual table rows instead of applied to the table as a whole, which
makes the responsive design more flexible.