feat: add serverside filter-by-date to logged records
Allows the option to filter FoodIntakeRecords and BodyWeightRecords by
date in frontend. Filtering is necessarily performed in the backend
to play nicely with pagination.
Also adds new component DateFilterDialog to handle UI.
The UI is still clunky (it should propbably be easier to see and clear
currently applied filter without having to go through filter dialog),
but I'm shipping for now to interact with it, to get a better idea of
how I would like to change the UI.